// lib/api/generated/models/flagAudience.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type FlagAudience = typeof FlagAudience[keyof typeof FlagAudience];


export const FlagAudience = {
  public: 'public',
  panels: 'panels',
  both: 'both',
  internal: 'internal',
} as const;